requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
python-dotenv
openai
fastapi
uvicorn
langgraph
langchain
langchain-openai
python-dotenv
faiss-cpu
numpy
openai
tavily-python
langfuse
pydantic
fastapi
uvicorn
gunicorn
python-dotenv
langchain
langgraph
langchain-openai
langchain-community
langchain-experimental
langchain-tavily
tavily-python
faiss-cpu
openai
pydantic